Split off the rans_byte.h portion of rANS_static.c back into its own file, for reasons of copyright clarity.
Fixed the arguments to rans_uncompress and indeed the whole notion of
having RANS0 and RANS1 as two distinct codecs.  This was historical.
They got merged into a genric RANS codec during the CRAM3 discussions.
