+++ cram/rANS_static.h
@@ -4,7 +4,7 @@
 unsigned char *rans_compress(unsigned char *in, unsigned int in_size,
 			     unsigned int *out_size, int order);
 unsigned char *rans_uncompress(unsigned char *in, unsigned int in_size,
-			       unsigned int *out_size, int order);
+			       unsigned int *out_size);
 
 
 #endif /* RANS_STATIC_H */
